Fixed: delete file
