+++ android/app/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.1" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="cc.seeed.iot" >
 
@@ -23,7 +23,7 @@
             android:theme="@style/Theme.NodeSet" >
         </activity>
 
-        <activity android:name=".ui_setup.SetupActivity" >
+        <activity android:name=".ui_login.SetupActivity" >
 
         </activity>
         <activity
