+++ android/app/src/main/java/cc/seeed/iot/ui_login/SignInDialogFragment.java
@@ -1,4 +1,4 @@
-package cc.seeed.iot.ui_setup;
+package cc.seeed.iot.ui_login;
 
 import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
@@ -14,7 +14,6 @@ import android.text.TextUtils;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.AutoCompleteTextView;
 import android.widget.Button;
 import android.widget.EditText;
@@ -31,23 +30,25 @@ import retrofit.Callback;
 import retrofit.RetrofitError;
 
 /**
- * Created by tenwong on 15/6/30.
+ * Created by tenwong on 15/7/1.
  */
-public class SignUpDialogFragment extends DialogFragment {
+public class SignInDialogFragment extends DialogFragment {
     Context context;
     User user;
 
-    AutoCompleteTextView mEmailView;
-    EditText mPasswordView;
-    EditText mPasswordVerifyView;
+    AlertDialog alertDialog;
+
+    private AutoCompleteTextView mEmailView;
+    private EditText mPasswordView;
+
     private View mProgressView;
-    private View mLoginRegisterView;
-    private AlertDialog alertDialog;
+    private View mLoginFormView;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         context = getActivity();
+
         user = ((MyApplication) getActivity().getApplication()).getUser();
     }
 
@@ -55,17 +56,16 @@ public class SignUpDialogFragment extends DialogFragment {
     public Dialog onCreateDialog(Bundle savedInstanceState) {
         AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
         LayoutInflater inflater = getActivity().getLayoutInflater();
-        View view = inflater.inflate(R.layout.sign_up, null);
+        View view = inflater.inflate(R.layout.dialog_sign_in, null);
 
         mEmailView = (AutoCompleteTextView) view.findViewById(R.id.email);
         mPasswordView = (EditText) view.findViewById(R.id.password);
-        mPasswordVerifyView = (EditText) view.findViewById(R.id.verify);
         mProgressView = view.findViewById(R.id.login_progress);
-        mLoginRegisterView = view.findViewById(R.id.email_register_form);
+        mLoginFormView = view.findViewById(R.id.email_login_form);
 
         builder.setView(view);
-        builder.setTitle("Sign Up");
-        builder.setPositiveButton("Sign up", null);
+        builder.setTitle("Sign In");
+        builder.setPositiveButton("Sign In", null);
         builder.setNegativeButton("Cancel", null);
 
         return builder.create();
@@ -73,7 +73,7 @@ public class SignUpDialogFragment extends DialogFragment {
 
     @Override
     public void onStart() {
-        super.onStart();
+        super.onStart();    //super.onStart() is where dialog.show() is actually called on the underlying dialog, so we have to do it after this point
         alertDialog = (AlertDialog) getDialog();
         if (alertDialog != null) {
             Button positiveButton = (Button) alertDialog.getButton(Dialog.BUTTON_POSITIVE);
@@ -82,7 +82,7 @@ public class SignUpDialogFragment extends DialogFragment {
                 public void onClick(View v) {
                     Boolean wantToCloseDialog = false;
                     //Do stuff, possibly set wantToCloseDialog to true then...
-                    attemptRegister();
+                    attemptLogin();
                     if (wantToCloseDialog)
                         alertDialog.dismiss();
                     //else dialog stays open. Make sure you have an obvious way to close the dialog especially if you set cancellable to false.
@@ -91,29 +91,22 @@ public class SignUpDialogFragment extends DialogFragment {
         }
     }
 
-    private void attemptRegister() {
+    private void attemptLogin() {
         mEmailView.setError(null);
         mPasswordView.setError(null);
 
         String email = mEmailView.getText().toString();
         String password = mPasswordView.getText().toString();
-        String passwordVerify = mPasswordVerifyView.getText().toString();
 
         boolean cancel = false;
         View focusView = null;
-
-        if (!(password.equals(passwordVerify))) {
-            mPasswordVerifyView.setError("not same password");
-            focusView = mPasswordVerifyView;
-            cancel = true;
-        }
 
         if (TextUtils.isEmpty(password) || !isPasswordValid(password)) {
             mPasswordView.setError("invalid Password");
             focusView = mPasswordView;
             cancel = true;
+            Log.e("iot", "password error");
         }
-
 
         if (TextUtils.isEmpty(email)) {
             mEmailView.setError("Require email");
@@ -121,7 +114,7 @@ public class SignUpDialogFragment extends DialogFragment {
             cancel = true;
 
         } else if (!isEmailValid(email)) {
-            mEmailView.setError("Invalid email");
+            mEmailView.setError("Error email");
             focusView = mEmailView;
             cancel = true;
         }
@@ -133,7 +126,7 @@ public class SignUpDialogFragment extends DialogFragment {
             final String fianlEmail = email;
             IotApi api = new IotApi();
             IotService iot = api.getService();
-            iot.userCreate(email, password, new Callback<UserResponse>() {
+            iot.userLogin(email, password, new Callback<UserResponse>() {
                 @Override
                 public void success(UserResponse userResponse, retrofit.client.Response response) {
                     String status = userResponse.status;
@@ -162,14 +155,6 @@ public class SignUpDialogFragment extends DialogFragment {
         }
     }
 
-    private boolean isEmailValid(String email) {
-        return email.contains("@");
-    }
-
-    private boolean isPasswordValid(String password) {
-        return password.length() >= 6;
-    }
-
 
     /**
      * Shows the progress UI and hides the login form.
@@ -182,12 +167,12 @@ public class SignUpDialogFragment extends DialogFragment {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB_MR2) {
             int shortAnimTime = getResources().getInteger(android.R.integer.config_shortAnimTime);
 
-            mLoginRegisterView.setVisibility(show ? View.GONE : View.VISIBLE);
-            mLoginRegisterView.animate().setDuration(shortAnimTime).alpha(
+            mLoginFormView.setVisibility(show ? View.GONE : View.VISIBLE);
+            mLoginFormView.animate().setDuration(shortAnimTime).alpha(
                     show ? 0 : 1).setListener(new AnimatorListenerAdapter() {
                 @Override
                 public void onAnimationEnd(Animator animation) {
-                    mLoginRegisterView.setVisibility(show ? View.GONE : View.VISIBLE);
+                    mLoginFormView.setVisibility(show ? View.GONE : View.VISIBLE);
                 }
             });
 
@@ -203,7 +188,17 @@ public class SignUpDialogFragment extends DialogFragment {
             // The ViewPropertyAnimator APIs are not available, so simply show
             // and hide the relevant UI components.
             mProgressView.setVisibility(show ? View.VISIBLE : View.GONE);
-            mLoginRegisterView.setVisibility(show ? View.GONE : View.VISIBLE);
+            mLoginFormView.setVisibility(show ? View.GONE : View.VISIBLE);
         }
     }
+
+    private boolean isEmailValid(String email) {
+        return email.contains("@");
+    }
+
+    private boolean isPasswordValid(String password) {
+        return password.length() >= 6;
+    }
 }
+
+
