+++ android/app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
         android:id="@+id/pager"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="cc.seeed.iot.ui_setup.SetupActivity" />
+        tools:context="cc.seeed.iot.ui_login.SetupActivity" />
 
     <com.viewpagerindicator.CirclePageIndicator
         android:id="@+id/indicator"
