+++ android/app/src/main/res/layout/dialog_sign_in.xml
