+++ android/app/src/main/res/layout/dialog_sign_up.xml
