+++ android/app/src/main/res/layout/smartconfig_connect_backup.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="cc.seeed.iot.ui_setup.test">
+    tools:context="cc.seeed.iot.ui_login.test">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
