+++ android/app/src/main/res/layout/smartconfig_connected.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:background="@color/apptheme_blue"
     android:orientation="vertical"
-    tools:context="cc.seeed.iot.ui_setup.test">
+    tools:context="cc.seeed.iot.ui_login.test">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
